requirements.txt
# Kraken Momentum Trader - Dependencies

# HTTP Requests
requests>=2.31.0

# Configuration
pyyaml>=6.0.1

# Environment Variables
python-dotenv>=1.0.0

# AWS SDK (for production deployment)
boto3>=1.34.0

# CloudWatch Logging (optional, for AWS)
watchtower>=3.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Type Checking (development)
mypy>=1.8.0

# Code Formatting (development)
black>=24.0.0
isort>=5.13.0

# Linting (development)
flake8>=7.0.0
